requirements.txt
Django>=3.2.3,<3.3
psycopg2>=2.8.6,<2.9app
djangorestframework
uWSGI>=2.0.19.1,<2.1